# apps/api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: AGRORED API Gateway
  version: 0.1.0
  description: Punto de entrada del ecosistema AGRORED para enrutar peticiones a los microservicios.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRegisterPayload:
      type: object
      required: [tenantId, email, fullName, role, password]
      properties:
        tenantId:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
          minLength: 3
        role:
          type: string
          enum: [admin_municipal, producer, supermarket, community_kitchen, logistics_operator, territorial_analyst]
        password:
          type: string
          minLength: 8
    LoginPayload:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
security:
  - BearerAuth: []
paths:
...
  /api/v1/users/register:
    post:
      summary: Registrar usuario
      security: []
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterPayload"
      responses:
        "201":
          description: Usuario registrado
        "400":
          description: Payload invalido
        "409":
          description: Email ya existe
  /api/v1/users/login:
    post:
      summary: Iniciar sesion
      security: []
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginPayload"
      responses:
        "200":
          description: JWT token
        "401":
          description: Credenciales invalidas
  /api/v1/users/recover-password:
    post:
      summary: Solicitar recuperacion de contrasena
      security: []
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Mensaje generico para evitar enumeracion de correos
  /api/v1/users/reset-password:
    post:
      summary: Restablecer contrasena con token de recuperacion
      security: []
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Contrasena restablecida
        "400":
          description: Token invalido o payload incorrecto
...
  /api/v1/offers:
    get:
      summary: Listar ofertas
      tags: [Offers]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de ofertas
  /api/v1/offers/{id}:
    get:
      summary: Obtener oferta por ID
      tags: [Offers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle de la oferta
        "404":
          description: No encontrado
...
  /api/v1/rescues:
    get:
      summary: Listar rescates
      tags: [Rescues]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de rescates
  /api/v1/rescues/{id}:
    get:
      summary: Obtener rescate por ID
      tags: [Rescues]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle del rescate
        "404":
          description: No encontrado
  /api/v1/demands/register:
    post:
      summary: Registrar demanda institucional
      tags: [Demands]
      responses:
        "201":
          description: Demanda registrada
        "400":
          description: Payload invalido
  /api/v1/demands:
    get:
      summary: Listar demandas
      tags: [Demands]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de demandas
  /api/v1/demands/{id}:
    get:
      summary: Obtener demanda por ID
      tags: [Demands]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle de la demanda
        "404":
          description: No encontrado
...
  /api/v1/inventory:
    get:
      summary: Listar inventario
      tags: [Inventory]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de inventario
  /api/v1/inventory/{id}:
    get:
      summary: Obtener item de inventario por ID
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle del inventario
        "404":
          description: No encontrado
...
  /api/v1/logistics:
    get:
      summary: Listar ordenes logisticas
      tags: [Logistics]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de ordenes
  /api/v1/logistics/{id}:
    get:
      summary: Obtener orden logistica por ID
      tags: [Logistics]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle de la orden
        "404":
          description: No encontrado
...
  /api/v1/incidents:
    get:
      summary: Listar incidencias
      tags: [Incidents]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de incidencias
  /api/v1/incidents/{id}:
    get:
      summary: Obtener incidencia por ID
      tags: [Incidents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle de la incidencia
        "404":
          description: No encontrado
...
  /api/v1/audit:
    get:
      summary: Consultar log de auditoria (admin)
      tags: [Audit]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: Log de auditoria paginado
        "403":
          description: Sin permisos